The YKncok Filter

This function runs the Knockoffs procedure from start to finish, selecting variables relevant for predicting the outcome of interest.

YKnock_filter(
  X,
  Y,
  knockoffs = create_YKnock,
  statistic = stat_modelY_coef,
  fdr = 0.1,
  offset = 1
)

Arguments

X

n-by-p matrix or data frame of predictors.

Y

n-by-r matrix or data frame of responses.

knockoffs

method used to construct knockoffs for the \(X\) variables. It must be a function taking a n-by-p matrix as input and returning a n-by-p matrix of knockoff variables. By default, approximate model-X Gaussian knockoffs are used.

statistic

statistics used to assess variable importance. By default, a lasso statistic with cross-validation is used. See the Details section for more information.

fdr

target false discovery rate (default: 0.1).

offset

either 0 or 1 (default: 1). This is the offset used to compute the rejection threshold on the statistics. The value 1 yields a slightly more conservative procedure ("knockoffs+") that controls the false discovery rate (FDR) according to the usual definition, while an offset of 0 controls a modified FDR.

Value

An object of class "knockoff.result". This object is a list containing at least the following components:

Y

matrix of original responses

Yk

matrix of knockoff responses

statistic

computed test statistics

threshold

computed selection threshold

selected

named vector of selected variables

Details

This function creates the knockoffs, computes the importance statistics, and selects variables. It is the main entry point for the knockoff package.

The parameter knockoffs controls how knockoff variables are created. By default, the model-Y scenario is assumed and a multivariate normal distribution is fitted to the original variables \(Y\).

The default importance statistic is stat_modelY_coef.

It is possible to provide custom functions for the knockoff constructions or the importance statistics. Some examples can be found in the vignette.

References

Identification of Significant Gene Expression Changes in Multiple Perturbation Experiments using Knockoffs Tingting Zhao, Guangyu Zhu, Patrick Flaherty bioRxiv 2021.10.18.464822; https://www.biorxiv.org/content/10.1101/2021.10.18.464822v1

Examples

r = 100;p = 20;n = 40
m = 10  # num of important response
rho=0.2
betaValue = 1.5
SigmaX=matrix(rho,p,p)
diag(SigmaX)=1
betaA=matrix(sample(c(-1,1)*betaValue,size=m*p, replace = TRUE), nrow = m,ncol = p)
beta=matrix(0,r,p)
beta[1:m,]=betaA
sigma=1
X = matrix(rnorm(n*p),n)%*%chol(SigmaX)
Y = X %*% t(beta) + sqrt(sigma)*matrix(rnorm(n*r),n,r)
result=YKnock_filter(X, Y)
print(result$selected)
#>  [1]   1   2   3   4   5   6   7   8   9  10  15  20  22  23  34  37  38  48  52
#> [20]  54  66  68  78  80  83  84  94  98 102 104 105 109 110 114 115 119 138 141
#> [39] 142 148 152 159 165 168 170 174 178 190

# Basic usage with default arguments
result = YKnock_filter (X, Y)
print(result$selected)
#>  [1]   1   2   3   4   5   6   7   8   9  10  12  15  20  22  25  28  34  37  38
#> [20]  41  48  52  54  66  78  80  83  94  98 105 106 115 118 121 122 123 127 129
#> [39] 136 139 140 142 149 151 152 153 157 160 165 166 167 168 182 195 198